Update comments.php
Forgot to escape the $post string for comments.php, and the "append hashtags" function wasn't working whenever an image was added to the post, since the image had quotes in it. Obviously would have been lots of other problems, too, probably including SQL injection attacks. The programmers of this website are stoners, so we need all the help we can get to catch these kinds of little (big) mistakes!