Update header.php
Takes the header image and slogan from the top of every page and puts it into header.php to consolidate it into one place.
Also, added links to other p0rtals that are set up to go to the corresponding url for the p0rtal you link to. This would allow someone to search the same thing on many different p0rtals very easily, for instance. Or switch between all of their about pages.

Now all the pages must be updated to not have that top html. And the new style.css needs to be added for "link2p0rtals" id to be styled properly.