Update hashtaglib.php
Added a requirement for a space right before hashtag, since that allows urls to have a # in them without conflict. Also, simulates a space in at the beginning of any string it's asked to assess, as a shortcut to fix the problem of hashtags not being acknowledged if they are at the very beginning of the string.